requirements.txt
flask==2.1.2
flask_restful==0.3.9
flask_migrate==3.1.0
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.3
requests==2.27.1